refactor: merge splittered shared specs
