+++ specs/processors/can-be.spec.ts
@@ -1,6 +1,6 @@
 import { build, createBuildable } from '../../src/core';
 import { canBe } from '../../src/processors/can-be';
-import { includeProcessorFnSpecs } from './shared/shared-specs';
+import { includeProcessorFnSpecs } from '../spec-helpers/shared-specs';
 
 describe('canBe processor function', () => {
   it('should replace the value of a node with the specified one', () => {
